framework: Add RIL for qualcom basebands with no sim status I will describe the behaviour below because I think it may be useful for others with updated qualcomm basebands.
Semc recently upgraded to a new baseband version which changes
the way SIM status works. At the core of this changs is the fact
that basebands are now supporting multiple SIM cards and this means
that sending the app state as an extra of
RIL_UNSOL_RESPONSE_RADIO_STATE_CHANGED

CAF supports this behaviour in the gingerbread_rel branch but
the changes are unmergeable due to other complex that would break
all devices.

This is a workaround where the new state (radio status) is parsed
from RIL_UNSOL_RESPONSE_RADIO_STATE_CHANGED and this triggers a
thread which will in turn query the SIM card and send SIM/RUIM_READY
when appropriate.

Revised by Giulio Cervera for ICS

Change-Id: I5bbd10ec5bd0e4f6c4791def905fe5677ba8f690