frameworks/base: wifi: Immediate disconnect in idle state
After the idle period when wifi is disconnected, wifi state machine
waits for an additional 2 mintues when it is not in emergency
mode.  However since 'Handler' is used without a wake lock, phone
can go to suspend mode. So if it is an idle disconnect, disconnect
immediately without doing the delayed stop.

Change-Id: I79499db07569841386a1587ec6d89424fb47d698
CRs-Fixed: 329183