LGEStarRIL: fix DST handling exception (ported from GB)
current implementation of NITZ assumes that it is in the form of OFFSET,"DATE,TIME,DST"
and split it into OFFSET and "DATE,TIME" and "DST" and assign DST to parcelextra.
some carriers send NITZ without DST flag, string parsing splits DATE and TIME
and assign TIME for DST, which leads to throwing integer parsing exception,
and then possibly RIL crash
this commit utilize String.split(",") to split NITZ and reconstruct it correctly,
and handle DST as optional.
(also moved offset evaluation to inside only where it is needed.)

Change-Id: I50a8462bf1e2ea6820da4a006fd12ced8d224446