Framework: Use ICS alert drawable
Merging the reboot menu code caused a removal of
the bugfix in this commit: http://bit.ly/MYAIxj

ICS updated the way one should point to the alert
drawable. The old way makes it point to an upscaled
GB drawable instead.

NOTE: Google actually removed the icon for the
shutdown menu in the commit below but I think
it looks nice.

Change-Id: If37c68507886607f9be7f733c0586f814f0ed682
http://bit.ly/LJvp3h