libstagefright: Added suspend-resume support for 3D clips
On suspend-resume, 3D clips were not being detected as 3D and were being
played on 2D.  As we never send the frame with 3D SEI (typically the first
frame) on resume, it is necessary to store state in between.  Due to the
loss of the suspend-resume APIs in stagefright, mediaplayer can no longer
store state between s/r.  As a result, we have to make VideoView hold state.

This change sends a notification (MEDIA_INFO_3D_ATTRIBUTES)to the upper layer
with the appropriate 3D flags.  On resume, the upper layer is expected
to pass these flags back to stagefright via the setParameters call. This
in turn sets the appropriate key in metadata. In processSEI, if we
happen to see this key, we pass the value associated with the key to
NativeWindow.

CRs-Fixed: 355006
(cherry picked from commit 931ab52420061c10120d658058e9926e75f6338f)

Change-Id: I08f0c8fbf4febc6a6dbc1d2f4125b2d59fbb6e42

Conflicts:

	media/libstagefright/OMXCodec.cpp